requirements.txt
bcrypt==3.2.0
celery==5.1.2
Django==3.2.5
django-crispy-forms==1.12.0
Pillow==8.3.1
psycopg2==2.9.1
redis==3.5.3
requests==2.26.0
